requirements.txt

pysqlite>=3.19
tweepy>=3.5.0
pyyaml>=3.12
